refactor(wiki): relocate wiki domain models to app/models/wiki.py
Moves ChangeKind, CommitResult, and AiRebaseMaxRetriesError from
app/wiki/models.py to app/models/wiki.py, consistent with the
app/models/ layout (file_system.py, trigger.py, etc.).
